requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==3.2.13
django-environ==0.9.0
djangorestframework==3.14.0
flake8==6.0.0
idna==3.4
isort==5.12.0
mccabe==0.7.0
pycodestyle==2.10.0
pyflakes==3.0.1
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
stripe==5.1.1
urllib3==1.26.14
